Split out operations functionality, rename EnqueueV2
