+++ cmd/juju/action/action.go
@@ -29,10 +29,10 @@ type APIClient interface {
 	// TODO(juju3) - remove.
 	Enqueue(params.Actions) (params.ActionResults, error)
 
-	// EnqueueV2 takes a list of Actions and queues them up to be executed by
-	// the designated ActionReceiver, returning the ids of the overall operation
-	// and each individual task.
-	EnqueueV2(params.Actions) (params.EnqueuedActions, error)
+	// EnqueueOperation takes a list of Actions and queues them up to be executed as
+	// an operation, each action running as a task on the the designated ActionReceiver.
+	// We return the ID of the overall operation and each individual task.
+	EnqueueOperation(params.Actions) (params.EnqueuedActions, error)
 
 	// ListAll takes a list of Tags representing ActionReceivers and returns
 	// all of the Actions that have been queued or run by each of those
